Deduplicate stage definition
